requirements.txt
# Core dependencies for offline machine defect detection
# Install with: pip install -r requirements.txt

# Vision and ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.25.0
sentencepiece>=0.1.99
pillow>=10.0.0

# Camera and image processing
opencv-python>=4.8.0
numpy>=1.24.0

# Hugging Face hub (for model download - run once while online)
huggingface-hub>=0.19.0

# Optional: quantization support
bitsandbytes>=0.41.0
optimum>=1.14.0

# Utils
python-dotenv>=1.0.0
flask>=3.0.0
